+++ src/pages/employee/[id]/index.vue
@@ -32,6 +32,8 @@ const profilePictureToUpload = ref<File>()
 const updateProfilePicturePopup = ref(false)
 const keyIncrement = ref(1)
 const loading = ref(false)
+const deleteLoading = ref(false)
+const uploadLoading = ref(false)
 const employeeForm = useEmployeeForm()
 
 const notif = useNotyf()
@@ -123,6 +125,7 @@ const fetchCurrentProfilePic = async () => {
     await sleep(500)
     if (profile_pic.media.length != 0) {
         employeeProfilePicture.value = profile_pic.media[profile_pic.media.length - 1]
+
     }
 
 }
@@ -158,7 +161,7 @@ const onAddFile = async (event: any) => {
 
 }
 const UploadFile = async () => {
-
+    uploadLoading.value = true
     let formData = new FormData();
     if (filesToUpload.value != undefined)
         formData.append('images[]', filesToUpload.value);
@@ -174,12 +177,15 @@ const UploadFile = async () => {
         media[0].relative_path = import.meta.env.VITE_MEDIA_BASE_URL + media[0].relative_path
 
         employeeFiles.value.push(media[0])
+        filesToUpload.value = undefined
+        uploadLoading.value = false
         return true
     }
     else {
         // @ts-ignore
         await sleep(200);
-
+        filesToUpload.value = undefined
+        uploadLoading.value = false
         notif.error(message)
 
     }
@@ -192,7 +198,7 @@ const onDeleteFile = (file_id: number) => {
 
 }
 const removefile = async () => {
-
+    deleteLoading.value = true
     const { message, success } = await deleteFile(deleteFileId.value)
 
     if (success) {
@@ -203,9 +209,13 @@ const removefile = async () => {
 
         employeeFiles.value.splice(employeeFiles.value.findIndex((element) => element.id == deleteFileId.value), 1)
         deleteFilePopup.value = false
+        deleteLoading.value = false
+
         return true
     }
     else {
+        deleteLoading.value = false
+
         // @ts-ignore
         await sleep(200);
 
@@ -253,9 +263,11 @@ const onEditProfilePicture = async (error: any, fileInfo: any) => {
     }
 }
 const UploadProfilePicture = async () => {
+    uploadLoading.value = true
+
     let _success = true
     let _message = ''
-    if (employeeProfilePicture.value.id != undefined) {
+    if (employeeProfilePicture.value.id != undefined && profilePictureToUpload.value != undefined) {
         const { message, success } = await deleteFile(employeeProfilePicture.value.id)
         _success = success
         _message = message
@@ -281,6 +293,9 @@ const UploadProfilePicture = async () => {
 
             keyIncrement.value++
             updateProfilePicturePopup.value = false
+            profilePictureToUpload.value = undefined
+            uploadLoading.value = false
+
 
         }
         else {
@@ -289,6 +304,8 @@ const UploadProfilePicture = async () => {
 
             notif.error(message)
             keyIncrement.value++
+            uploadLoading.value = false
+            profilePictureToUpload.value = undefined
 
             updateProfilePicturePopup.value = false
 
@@ -296,9 +313,11 @@ const UploadProfilePicture = async () => {
     } else {
         // @ts-ignore
         await sleep(200);
+        uploadLoading.value = false
 
         notif.error(_message)
         keyIncrement.value++
+        profilePictureToUpload.value = undefined
 
         updateProfilePicturePopup.value = false
 
@@ -308,6 +327,7 @@ const UploadProfilePicture = async () => {
 }
 const RemoveProfilePicture = async () => {
     if (employeeProfilePicture.value.id != undefined) {
+        deleteLoading.value = true
         const { message, success } = await deleteFile(employeeProfilePicture.value.id)
         if (success) {
             await sleep(200);
@@ -321,7 +341,9 @@ const RemoveProfilePicture = async () => {
             notif.error(message)
 
         }
+        profilePictureToUpload.value = undefined
         keyIncrement.value++
+        deleteLoading.value = false
 
         updateProfilePicturePopup.value = false
 
@@ -550,10 +572,12 @@ const RemoveProfilePicture = async () => {
                                                 </div>
                                             </VControl>
                                         </VField>
-                                        <VButton v-if="filesToUpload != undefined" @click="UploadFile" class=""
-                                            icon="lnir lnir-add-files rem-100" light dark-outlined>
-                                            Upload
-                                        </VButton>
+                                        <VLoader size="small" :active="uploadLoading">
+                                            <VButton v-if="filesToUpload != undefined" @click="UploadFile" class=""
+                                                icon="lnir lnir-add-files rem-100" light dark-outlined>
+                                                Upload
+                                            </VButton>
+                                        </VLoader>
                                     </div>
                                     <h6 class="ml-2 mt-2 help">Max size: 2 MB | Accepted file types :
                                         png,
@@ -635,7 +659,10 @@ const RemoveProfilePicture = async () => {
             <VPlaceholderSection title="Are you sure?" :subtitle="`you are about to delete this file permenantly`" />
         </template>
         <template #action="{ close }">
-            <VButton color="primary" raised @click="removefile()">Confirm</VButton>
+            <VLoader size="small" :active="deleteLoading">
+
+                <VButton color="primary" raised @click="removefile()">Confirm</VButton>
+            </VLoader>
         </template>
     </VModal>
     <VModal :key="keyIncrement" title="Update Profile Picture" :open="updateProfilePicturePopup" actions="center"
@@ -662,9 +689,17 @@ const RemoveProfilePicture = async () => {
         </template>
 
         <template #action="{ close }">
-            <VButton v-if="employeeProfilePicture.id != undefined" color="warning" raised @click="RemoveProfilePicture">
-                Delete</VButton>
-            <VButton color="primary" raised @click="UploadProfilePicture">Update</VButton>
+            <VLoader size="small" :active="deleteLoading">
+
+                <VButton v-if="employeeProfilePicture.id != undefined" color="danger" outlined class="mr-2"
+                    @click="RemoveProfilePicture">
+                    Delete</VButton>
+            </VLoader>
+            <VLoader size="small" :active="uploadLoading">
+
+                <VButton color="primary" raised @click="UploadProfilePicture">Update</VButton>
+            </VLoader>
+
         </template>
     </VModal>
